% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_clean_df.R
\name{build_clean_df}
\alias{build_clean_df}
\title{Reproduces the output of 'clean.csv'}
\usage{
build_clean_df(df_input)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
a clean csv
}
\description{
Computes all columns of interest, with a catch-except so everything does not fail at once.
}
