% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_blau_alter_heterophily.R
\name{calc_blau_alter_heterophily}
\alias{calc_blau_alter_heterophily}
\title{Calculate Blau heterophily index for a given alter attribute}
\usage{
calc_blau_alter_heterophily(persnet_row, attribute = NULL)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{attribute}{One of 'gender', 'educ',
'distance' (how far alters live from ego), 'length' (how long alters have
known ego), or 'speak' (how often alters speak to ego).}
}
\value{
Blau heterophily index for the specified attribute.
}
\description{
Computes the Blau heterophily index for a specified alter attribute.
The index measures diversity within a personal network.
n.b. as alters can be of multiple 'race' and offer several 'type of support', code is included
that can be modified/used if appropriate for users. Blau heterophily not defined for overlaping categories
}
