% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ego_constraint_df.R
\name{calc_ego_constraint_df}
\alias{calc_ego_constraint_df}
\title{Calculate Ego Constraint from a persnet dataframe}
\usage{
calc_ego_constraint_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A named vector containing the ego's network constraint for each network.
}
\description{
Computes the ego network constraint for a persnet dataframe.
}
