% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ego_ens_df.R
\name{calc_ego_ens_df}
\alias{calc_ego_ens_df}
\title{Calculate Ego Effective Network Size (ENS) from a persnet dataframe}
\usage{
calc_ego_ens_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A named vector containing the effective network size (ENS) for each node in the network.
}
\description{
Computes the effective network size (ENS) for a persnet dataframe
}
