% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_density.R
\name{calc_egoless_density}
\alias{calc_egoless_density}
\title{Calculate the density of a network after removing the ego node}
\usage{
calc_egoless_density(tg_graph)
}
\arguments{
\item{tg_graph}{A tidygraph object representing a personal network.}
}
\value{
The density of the network without the ego node.
}
\description{
Computes the density of a personal network graph after removing the ego node.
}
