% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_density_df.R
\name{calc_egoless_density_df}
\alias{calc_egoless_density_df}
\title{Calculate Egoless Density from a persnet dataframe}
\usage{
calc_egoless_density_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A vector containing the egoless density for each network.
}
\description{
Computes the density of each personal network after removing the ego node from each network 
contained in a persnet dataframe.
}
\examples{
# Assuming persnet_df is your personal network dataframe:
vector_egoless_density <- calc_egoless_density_df(persnet_df)
}
