% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_max_degree_df.R
\name{calc_egoless_max_degree_df}
\alias{calc_egoless_max_degree_df}
\title{Calculate max degree}
\usage{
calc_egoless_max_degree_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A vector containing the egoless maximum degree for each network.
}
\description{
Computes the maximum degree (i.e., the highest number of connections)
for each personal network after removing the ego node from the network,
using a persnet dataframe as input.
}
