% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_mean_degree.R
\name{calc_egoless_mean_degree}
\alias{calc_egoless_mean_degree}
\title{Calculate the mean degree of a network after removing the ego node}
\usage{
calc_egoless_mean_degree(tg_graph)
}
\arguments{
\item{tg_graph}{A tidygraph object of a personal network}
}
\value{
The mean degree of nodes in the network (excluding ego).
        Returns 0 if the graph is empty and NA if an error occurs.
}
\description{
Calculate the mean degree of a network after removing the ego node
}
\examples{
calc_egoless_mean_degree(tg_graph)
}
