% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_min_degree.R
\name{calc_egoless_min_degree}
\alias{calc_egoless_min_degree}
\title{Calculate the minimum degree of a network after removing the ego node}
\usage{
calc_egoless_min_degree(tg_graph)
}
\arguments{
\item{tg_graph}{A tidygraph object representing a personal network.}
}
\value{
The minimum degree of nodes in the network (excluding ego).
        Returns 0 if the graph is empty and NA if an error occurs.
}
\description{
Computes the minimum degree of nodes in a personal network after removing the ego node.
}
\examples{
calc_egoless_min_degree(tg_graph)
}
