% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_egoless_min_degree_df.R
\name{calc_egoless_min_degree_df}
\alias{calc_egoless_min_degree_df}
\title{Calculate egoless minimum degree from a persnet dataframe}
\usage{
calc_egoless_min_degree_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A vector containing the egoless minimum degree for each network.
}
\description{
Computes the minimum degree (i.e., the lowest number of connections)
for each personal network after removing the ego node, using a persnet dataframe as input.
}
