% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_constraint.R
\name{calc_node_constraint}
\alias{calc_node_constraint}
\title{Calculate Burt's Constraint for a given node in a network}
\usage{
calc_node_constraint(tidygra, node_index = NULL)
}
\arguments{
\item{tidygra}{A tidygraph object representing a personal network.}

\item{node_index}{(Optional) The node for which constraint is calculated (default = "ego").}
}
\value{
The constraint value for the specified node.
        Returns NA if the network is an isolate.
}
\description{
Computes Burt's Constraint measure for a specified node in a tidygraph object.
Burt's Constraint quantifies how much a node’s connections are concentrated within
a closed group, limiting access to diverse information.
}
\examples{
calc_node_constraint(tidygra, "ego")
calc_node_constraint(tidygra, "burt")
}
\references{
Everett, M. G., & Borgatti, S. P. (2020). *Unpacking Burt’s constraint measure*.
Social Networks, 62, 50-57. ISSN 0378-8733.
\doi{10.1016/j.socnet.2020.02.001}
}
