% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_ens.R
\name{calc_node_ens}
\alias{calc_node_ens}
\title{Calculate the Effective Network Size (ENS) for a given node in a network}
\usage{
calc_node_ens(tidygra, node_index = NULL)
}
\arguments{
\item{tidygra}{A tidygraph object representing a personal network.}

\item{node_index}{(Optional) The node for which ENS is calculated (default = "ego").}
}
\value{
The Effective Network Size (ENS) value for the specified node.
        Returns NA if the network is an isolate.
}
\description{
Computes the Effective Network Size (ENS) for a specified node in a tidygraph object,
considering tie strengths. ENS is a measure of structural holes in a network.
}
\examples{
calc_node_ens(tidygra, "ego")
calc_node_ens(tidygra, "burt")
}
\references{
Burt, R. S. (1995). *Structural Holes: The Social Structure of Competition*.
Harvard University Press.
}
