% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_alcohol.R
\name{calc_prop_alters_alcohol}
\alias{calc_prop_alters_alcohol}
\title{Calculate the proportion of alters who do not drink heavily or whose drinking habits are unknown}
\usage{
calc_prop_alters_alcohol(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of alters who do not drink heavily or whose drinking habits are unknown.
}
\description{
Computes the proportion of alters who either do not drink heavily or whose drinking habits are unknown.
}
\examples{
calc_prop_alters_alcohol(persnet_row)
}
