% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_diet.R
\name{calc_prop_alters_diet}
\alias{calc_prop_alters_diet}
\title{Calculate the proportion of alters with a good diet}
\usage{
calc_prop_alters_diet(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of alters who `ate a healthy diet regularly over the past 3 months'
}
\description{
Computes the proportion of alters who have a good diet.
}
