% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_distance_away.R
\name{calc_prop_alters_distance_away}
\alias{calc_prop_alters_distance_away}
\title{Calculate the proportion of alters living at a specified distance from ego}
\usage{
calc_prop_alters_distance_away(persnet_row, distance_category)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{distance_category}{One of 'same_house', '1_5miles', '6_15miles', '16_50miles', or 'more50miles'.}
}
\value{
The proportion of alters within the specified distance category.
}
\description{
Computes the proportion of alters who live within a specified distance category from the ego.
}
\examples{
calc_prop_alters_distance_away(persnet_row, "same_house")
calc_prop_alters_distance_away(persnet_row, "1_5miles")
calc_prop_alters_distance_away(persnet_row, "6_15miles")
calc_prop_alters_distance_away(persnet_row, "16_50miles")
calc_prop_alters_distance_away(persnet_row, "more50miles")
}
