% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_education_level.R
\name{calc_prop_alters_education_level}
\alias{calc_prop_alters_education_level}
\title{Calculate the proportion of alters based on their education level}
\usage{
calc_prop_alters_education_level(persnet_row, education_level)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{education_level}{One of 'only_high_school', 'some_college', 'college_grad', or 'dont_know'.}
}
\value{
The proportion of alters within the specified education level.
}
\description{
Computes the proportion of alters who fall into a specified education level category.
}
\examples{
calc_prop_alters_education_level(persnet_row, "only_high_school")
calc_prop_alters_education_level(persnet_row, "some_college")
calc_prop_alters_education_level(persnet_row, "college_grad")
calc_prop_alters_education_level(persnet_row, "dont_know")
}
