% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_exercise.R
\name{calc_prop_alters_exercise}
\alias{calc_prop_alters_exercise}
\title{Calculate the proportion of alters who exercise regularly}
\usage{
calc_prop_alters_exercise(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of alters who exercise regularly.
}
\description{
Computes the proportion of alters who exercise regularly.
}
