% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_freq_speak.R
\name{calc_prop_alters_freq_speak}
\alias{calc_prop_alters_freq_speak}
\title{Calculate the proportion of alters who speak to ego at a specified frequency}
\usage{
calc_prop_alters_freq_speak(persnet_row, frequency)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{frequency}{One of 'daily', 'weekly', 'monthly', 'more_monthly', or 'dont_know'.}
}
\value{
The proportion of alters who speak to the ego at the specified frequency.
}
\description{
Computes the proportion of alters who speak to the ego at a specified frequency.
}
\examples{
calc_prop_alters_freq_speak(persnet_row, "daily")
calc_prop_alters_freq_speak(persnet_row, "weekly")
calc_prop_alters_freq_speak(persnet_row, "monthly")
calc_prop_alters_freq_speak(persnet_row, "more_monthly")
calc_prop_alters_freq_speak(persnet_row, "dont_know")
}
