% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_known_length.R
\name{calc_prop_alters_known_length}
\alias{calc_prop_alters_known_length}
\title{Calculate the proportion of alters based on the length of time known}
\usage{
calc_prop_alters_known_length(persnet_row, time_category)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{time_category}{One of 'less_than_3years', 'three_to_6years', 'more_than_6years', or 'unknown'.}
}
\value{
The proportion of alters within the specified time category.
}
\description{
Computes the proportion of alters who have known the ego for a specified length of time.
}
\examples{
calc_prop_alters_known_length(persnet_row, "less_than_3years")
calc_prop_alters_known_length(persnet_row, "three_to_6years")
calc_prop_alters_known_length(persnet_row, "more_than_6years")
calc_prop_alters_known_length(persnet_row, "unknown")
}
