% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_male_gender.R
\name{calc_prop_alters_male_gender}
\alias{calc_prop_alters_male_gender}
\title{Calculate the proportion of alters who are men}
\usage{
calc_prop_alters_male_gender(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of alters who are men.
}
\description{
Calculate the proportion of alters who are men
}
\examples{
calc_prop_alters_male_gender(persnet_row)
}
