% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_neg_health_influence.R
\name{calc_prop_alters_neg_health_influence}
\alias{calc_prop_alters_neg_health_influence}
\title{Proportion of alters with a negative influence on ego's health}
\usage{
calc_prop_alters_neg_health_influence(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of alters who negatively influence ego's health.
}
\description{
Calculates the proportion of alters in a personal network who negatively influence ego's health.
}
\examples{
calc_prop_alters_neg_health_influence(persnet_row)
}
