% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_race_identity.R
\name{calc_prop_alters_race_identity}
\alias{calc_prop_alters_race_identity}
\title{Proportion of alters by racial identity}
\usage{
calc_prop_alters_race_identity(persnet_row, race_category)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{race_category}{One of 'black', 'white', 'native_american', 'asian', 'pacific_islander', or 'unknown'.}
}
\value{
The proportion of alters within the specified racial category.
}
\description{
Calculates the proportion of alters in a personal network who belong to a specified racial category.
}
