% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_support.R
\name{calc_prop_alters_support}
\alias{calc_prop_alters_support}
\title{Proportion of alters providing a specific type of support}
\usage{
calc_prop_alters_support(persnet_row, support_type)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{support_type}{One of 'emotional', 'advice', 'finance', 'health', or 'camaraderie'.}
}
\value{
The proportion of alters providing the specified type of support.
        Returns NA if the network is an isolate.
}
\description{
Calculates the proportion of alters in a personal network who provide a specified type of support.
}
\examples{
calc_prop_alters_support(persnet_row, "emotional")
calc_prop_alters_support(persnet_row, "advice")
calc_prop_alters_support(persnet_row, "finance")
calc_prop_alters_support(persnet_row, "health")
calc_prop_alters_support(persnet_row, "camaraderie")
}
