% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_alters_df.R
\name{calc_total_alters_df}
\alias{calc_total_alters_df}
\title{Calculate Total Alters from a persnet dataframe}
\usage{
calc_total_alters_df(persnet_df)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A vector containing the total count of unique alters for each record.
}
\description{
Computes the total number of unique alters for each record in a persnet dataframe. This counts the number of unique strings in the `more names' columns and unique alters in the persnet tidgraph.
}
