% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_health_problems.R
\name{extract_health_problems}
\alias{extract_health_problems}
\title{Return the columns regarding health problems of the ego}
\usage{
extract_health_problems(persnet_row, health_position)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A string indicating the primary racial identification of the ego.
}
\description{
Return the columns regarding health problems of the ego
}
