% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_list_tidygraphs.R
\name{organize_list_tidygraphs}
\alias{organize_list_tidygraphs}
\title{Convert a persnet data frame into a list of tidygraph objects}
\usage{
organize_list_tidygraphs(persnet_df)
}
\arguments{
\item{persnet_df}{A persnet data frame.}
}
\value{
A list of tidygraph objects.
}
\description{
Converts a persnet data frame into a list of tidygraph objects, with each row 
representing an individual network.
}
\examples{
organize_list_tidygraphs(persnet_df)
}
