% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_networks.R
\name{plot_grid_networks}
\alias{plot_grid_networks}
\title{Plot a grid of personal network graphs}
\usage{
plot_grid_networks(list_tidygra, number_cols = NULL, nets_per_plot = NULL)
}
\arguments{
\item{list_tidygra}{A list of tidygraph objects.}

\item{number_cols}{Number of columns in the grid (default = 5).}

\item{nets_per_plot}{Number of networks per plot (default = 20).}
}
\value{
A combined grid plot of personal networks.
}
\description{
Visualizes multiple personal network graphs in a grid format using ggraph. 
Each individual network graph distinguishes between strong and weak ties and highlights the ego node.
}
\details{
- **Strong ties** are represented with solid red edges.
- **Weak ties** are represented with dashed blue edges.
- **Ego node** is represented with a black point.
- **Alters** are represented with grey points.

The function allows customization of the grid:
- **number_cols**: Specifies the number of columns in the grid (default = 5).
- **nets_per_plot**: Defines how many networks are displayed per plot (default = 20).
}
\examples{
plot_grid_networks(list_tidygra)
plot_grid_networks(list_tidygra, number_cols = 4, nets_per_plot = 10)
plot_grid_networks(list_tidygra[10:25], number_cols = 3, nets_per_plot = 15)
}
