% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_network.R
\name{plot_single_network}
\alias{plot_single_network}
\title{Plot a personal network graph using ggraph}
\usage{
plot_single_network(tidygra)
}
\arguments{
\item{tidygra}{A tidygraph object representing a personal network.}
}
\value{
A ggplot object visualizing the network structure.
        Returns NA if the input is invalid.
}
\description{
Visualizes a personal network graph using ggraph, distinguishing between strong and weak ties 
and highlighting the ego node. The plot uses different colors and line styles:
- **Strong ties** are represented with solid red edges.
- **Weak ties** are represented with dashed blue edges.
- **Ego node** is represented with a black point.
- **Alters** are represented with grey points.
}
