% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_heavy_drinkers_row.R
\name{prop_heavy_drinkers_row}
\alias{prop_heavy_drinkers_row}
\title{Calculate proportion of heavy drinkers in a persnet dataframe}
\usage{
prop_heavy_drinkers_row(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of heavy drinkers among the alters.
}
\description{
The function searches for alcohol-related alter columns (e.g., "name1alcohol" to "name15alcohol") and,
if found, calculates the ratio of alters with heavy drinking (indicated by values of 0 or 1)
to the total number of alters (as determined by tie columns). The result is rounded to two decimal places.
}
