% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_kin_persnet_row.R
\name{prop_kin_persnet_row}
\alias{prop_kin_persnet_row}
\title{Calculate Legacy Kin Proportion for a Personal Network Row}
\usage{
prop_kin_persnet_row(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network dataframe.}
}
\value{
A numeric value representing the legacy kin proportion for the given row.
}
\description{
Computes the legacy kin proportion for a given personal network row by summing the 
proportions of alters classified as "spouse" and "family".
}
