% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proportion_alters_neg_influence.R
\name{proportion_alters_neg_influence}
\alias{proportion_alters_neg_influence}
\title{Calculate proportion of alters ego answered had a negative influence}
\usage{
proportion_alters_neg_influence(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of alters who negatively impact ego
}
\description{
"Do any people in your network have a
negative in#uence on your health? For example, some people can passively or 
actively encourage you to smoke, not eat well, or not exercise."
}
