% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proportion_alters_provide_support.R
\name{proportion_alters_provide_support}
\alias{proportion_alters_provide_support}
\title{Calculate alters who provide some kind of support to ego}
\usage{
proportion_alters_provide_support(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of alters who "support ego most often"
}
\description{
Proportion of alters ego answered yes to "which person or persons support you most often?
You can choose more than one person."
}
