% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_propotion_alters_male.R
\name{proporton_alters_male}
\alias{proporton_alters_male}
\title{Calculate proportion of alters who are men}
\usage{
proporton_alters_male(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of alters who are men
}
\description{
Calculate proportion of alters who are men
}
