% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ego_from_tidygraph.R
\name{remove_ego_from_igraph}
\alias{remove_ego_from_igraph}
\title{Remove the ego node from a tidygraph object}
\usage{
remove_ego_from_igraph(tg_graph)
}
\arguments{
\item{tg_graph}{A tidygraph object representing a personal network.}
}
\value{
A tidygraph object without the ego node.
}
\description{
Removes the ego node from a tidygraph object, returning a graph without the ego.
This function is useful for analyzing the structure of an individual's network 
without the central ego node.
}
\examples{
remove_ego_from_igraph(tg_graph)
}
