% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_age_alters_row.R
\name{sd_age_alters_row}
\alias{sd_age_alters_row}
\title{Calculate standard deviation of Ages for a persnet dataframe row}
\usage{
sd_age_alters_row(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
A numeric value representing the standard deviation of the alters' ages,
        rounded to two decimal places.
}
\description{
Computes the standard deviation of the ages of the alters in an egocentric network
based on a single row of a persnet dataframe.
}
