% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_dur_prop_row.R
\name{weak_dur_prop_row}
\alias{weak_dur_prop_row}
\title{Calculate proportion of `weak' ties as defined as knowing ego less than six years}
\usage{
weak_dur_prop_row(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of alters who have known the ego for less than 6 years.
}
\description{
Computes the proportion of alters who have known the ego for less than 6 years in a persnet dataframe row.
This is done by summing the proportions of alters who have known the ego for "less_than_3years" and 
"three_to_6years", and then rounding the result to two decimal places.
}
