% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weak_freq_prop_row.R
\name{weak_freq_prop_row}
\alias{weak_freq_prop_row}
\title{Calculate Proportion of Weak Frequency Alters for a persnet dataframe row}
\usage{
weak_freq_prop_row(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe row}
}
\value{
A numeric value representing the proportion of alters with weak contact frequency (contacted monthly or more than monthly)
}
\description{
Computes the proportion of alters who contact the ego either "monthly" or "more_monthly"
from a single row of a personal network data frame, and rounds the result to two decimal places.
}
