% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_all_nodes_ens.R
\name{calc_all_nodes_ens}
\alias{calc_all_nodes_ens}
\title{Calculate Effective Network Size (ENS) for all nodes}
\usage{
calc_all_nodes_ens(tidygra)
}
\arguments{
\item{tidygra}{A tidygraph object representing a personal network}
}
\value{
A named vector containing ENS values for all nodes
}
\description{
Computes the Effective Network Size (ENS) for all nodes in a given
tidygraph object, returning a named vector of ENS values.
}
\references{
Burt, R. S. (1995). *Structural Holes: The Social Structure of Competition*.
Harvard University Press.
}
