% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attribute_iqv.R
\name{calc_attribute_iqv}
\alias{calc_attribute_iqv}
\title{Calculate the Index of Qualitative Variation (IQV) for a given alter attribute}
\usage{
calc_attribute_iqv(persnet_row, attribute = NULL)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}

\item{attribute}{One of `gender` or `educ`.}
}
\value{
The IQV value for the specified attribute.
}
\description{
Computes the Index of Qualitative Variation (IQV) for a specified alter attribute.
The IQV measures diversity and variation within a personal network.
}
