% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_ens.R
\name{calc_node_ens}
\alias{calc_node_ens}
\title{Calculate the Effective Network Size (ENS) for a given node in a network}
\usage{
calc_node_ens(tidygra, node_index = NULL)
}
\arguments{
\item{tidygra}{A tidygraph object representing a personal network.}

\item{node_index}{(Optional) The node for which ENS is calculated (default = "ego").}
}
\value{
The Effective Network Size (ENS), a numeric value, for the specified node.
        Returns NA if the network is an isolate.
}
\description{
Effective Network Size measures the number of non-redundant ties an ego node has. n.b.
This function treats missing weight values as equal to 1 and defines redundant tie in
terms of stong ties, not as a proportion of all ties as these networks are weighted.
ENS is calculated as:
\deqn{
  ENS_i = \sum_{j \neq i} \left[ 1 - \sum_{q \neq i,j} (p_{iq} \cdot m_{jq}) \right]
}
where:
\itemize{
  \item \eqn{i}: Ego node
  \item \eqn{j}: Direct neighbors (alters) of ego
  \item \eqn{q}: Other nodes in the network (excluding ego \eqn{i} and alter \eqn{j})
  \item \eqn{p_{iq}}: Proportion of ego's total tie strength directed to node \eqn{q}:
  \deqn{p_{iq} = \frac{w_{iq}}{\sum_{k \neq i} w_{ik}}}
  \item \eqn{m_{jq}}: Proportion of alter \eqn{j}'s strongest tie directed to node \eqn{q}:
  \deqn{m_{jq} = \frac{w_{jq}}{\max_{k \neq j}(w_{jk})}}
  \item \eqn{w_{ij}}: Weight of tie between node \eqn{i} and \eqn{j}
}
}
\examples{
\dontrun{
ens_value <- calc_node_ens(tidygra, node_index = "burt")
}
}
\references{
Burt, R. S. (1995). *Structural Holes: The Social Structure of Competition*.
Harvard University Press.
}
