% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_alcohol.R
\name{calc_prop_alters_alcohol}
\alias{calc_prop_alters_alcohol}
\title{Calculate the proportion of alters who who likely drank heavily in the past 3 months.}
\usage{
calc_prop_alters_alcohol(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of an ego's alters who likely drank heavily in the past 3 months.
}
\description{
Computes the proportion of alters who who likely drank heavily in the past 3 months.
}
\examples{
calc_prop_alters_alcohol(persnet_row)
}
