% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_alters_smoke.R
\name{calc_prop_alters_smoke}
\alias{calc_prop_alters_smoke}
\title{Proportion of alters who smoke}
\usage{
calc_prop_alters_smoke(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a personal network data frame.}
}
\value{
The proportion of an ego's alters who had likely smoked in the past 3 months.
}
\description{
Calculates the proportion of alters in a personal network who smoke.
}
\examples{
calc_prop_alters_smoke(persnet_row)
}
