% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_race_legacy.R
\name{extract_race_legacy}
\alias{extract_race_legacy}
\title{Return the race of the ego from the demographics questions}
\usage{
extract_race_legacy(persnet_row, race_position)
}
\arguments{
\item{persnet_df}{A dataframe containing personal network data.}
}
\value{
A string indicating the primary racial identification of the ego.
}
\description{
Returns the race of the ego from the demographics questions
}
