% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_networks.R
\name{plot_grid_networks}
\alias{plot_grid_networks}
\title{Plot a grid of personal network graphs}
\usage{
plot_grid_networks(list_tidygra, number_cols = NULL)
}
\arguments{
\item{list_tidygra}{A list of tidygraph objects.}

\item{number_cols}{Number of columns in the grid (default = ceiling of the square root of the total number of plots).}
}
\value{
A combined grid plot of sociograms
}
\description{
Visualizes multiple personal network graphs in a grid format using ggraph. 
Each individual network graph distinguishes between strong and weak ties and highlights the ego node.
}
\details{
- **Strong ties** are represented with solid red edges.
- **Weak ties** are represented with dashed blue edges.
- **Ego node** is represented with a black point.
- **Alters** are represented with grey points.

The function allows customization of the grid:
- **number_cols**: Specifies the number of columns in the grid 
(default = ceiling of the square root of the total number of plots to make).

Uses should subset the list networks fed into function to customize the number of plots.
Function not intended for visualizing more than 25 sociograms at a time.
}
\examples{
plot_grid_networks(list_tidygra)
plot_grid_networks(list_tidygra, number_cols = 4, nets_per_plot = 10)
plot_grid_networks(list_tidygra[10:25], number_cols = 3)
}
