% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_alters_provide_support.R
\name{proportion_alters_provide_support}
\alias{proportion_alters_provide_support}
\title{Calculate alters who provide some kind of support to ego}
\usage{
proportion_alters_provide_support(persnet_row)
}
\arguments{
\item{persnet_row}{A single row of a persnet dataframe.}
}
\value{
A numeric value representing the proportion of alters living more than 15 miles away.
}
\description{
Computes the proportion of alters who live more than 15 miles away from the ego in a persnet dataframe row.
This is done by summing the proportions of alters categorized as "16_50miles" and "more50miles", and rounding the result to two decimal places.
}
